PEtab

Documentation for PEtab.

PEtab.MeasurementData

PEtab.ParamData

PEtab.ParamMap

PEtab.ParameterIndices

PEtab.PeTabModel

PEtab.SimulationInfo

PEtab.XmlToModellingToolkit

PEtab.calcAccuracyOdeSolver

PEtab.calcCost

PEtab.calcGradCost!

PEtab.calcHessianApprox!

PEtab.calcHighAccOdeSolution

PEtab.calcLogLik

PEtab.calcLogLikNotSolveODE

PEtab.calcLogLikSolveODE

PEtab.calcSqErrVal

PEtab.changeExperimentalCond!

PEtab.changeModelParam!

PEtab.checkForPeTabFile

PEtab.createBigFloatODEProblem

PEtab.createBigFloatODEProblem

PEtab.createFileYmodSdU0

PEtab.createSdFunction

PEtab.createU0Function

PEtab.createYmodFunction

PEtab.getIndicesParam

PEtab.getNoiseParamStr

PEtab.getObsParamStr

PEtab.getRowExpId

PEtab.getSimulationInfo

PEtab.getTimeMax

PEtab.getWord

PEtab.isNumber

PEtab.peTabFormulaToJulia

PEtab.processMeasurementData

PEtab.processParameterData

PEtab.readDataFiles

PEtab.setParamToFileValues!

PEtab.setUpCostGradHess

PEtab.setUpPeTabModel

PEtab.solveOdeModelAllExperimentalCond!

PEtab.solveOdeModelAtFileValues

PEtab.solveOdeNoSS

PEtab.solveOdeSS

PEtab.transformObsOrData

PEtab.transformParamVec

PEtab.transformParamVec!

PEtab.transformYobsOrYmodArr!

PEtab.wordToJuliaSyntax

PEtab.writeODEModelToFile

PEtab.MeasurementData — Type

MeasurementData

Struct storing the data in the PeTab measurementData-file in type-stable manner.

Transform data supports log and log10 transformations of the data.

See also: [processMeasurementData]

PEtab.ParamData — Type

ParamData

Struct storing the data in the PeTab parameter-file in type-stable manner.

Currently logScale notices whether or not parameters are estimated on the log10 scale or not.

See also: [processParameterData]

PEtab.ParamMap — Type

ParamMap

Struct which makes out a map to correctly for an observation extract the correct observable or sd-param via the getObsOrSdParam function when computing the likelihood. Correctly built by buildMapParameters, and is part of the ParameterIndices-struct.

For noise or observable parameters belong to an observation, e.g (obsParam1, obsParam2), this struct stores which parameters should be estimtated, and for those parameters which index they correspond to in the parameter estimation vector. For constant parameters the struct stores the values.

See also: [getIndicesParam, buildMapParameters]

PEtab.ParameterIndices — Type

ParameterIndices

Struct storing names and mapping indices for mapping the parameter provided to the optimizers correctly.

Optimizers require a single vector input of parameters (pVecEst). However, the PeTab model has three kind of parameters, Dynmaic (part of the ODE-system), Observable (only part of the observation model) and Standard-deviation (only part of the standard deviation expression in the log-likelhood). This struct stores mapping indices (starting with i) to map pVecEst correctly when computing the likelihood (e.g map the SD-parameters in pVecEst correctly to a vector of SD-vals). Also stores the name of each parameter.

Furthermore, when computing yMod or SD the correct observable and sd parameters has to be used for each observation. The mapArrays effectively contains precomputed maps allowing said parameter to be effectively be extracted by the getObsOrSdParam function.

See also: [getIndicesParam, ParamMap]

PEtab.PeTabModel — Type

PeTabModel

Struct storing information about a PeTab-model. Create by the setUpCostFunc function.

Args

modelName: PeTab model name (must match the xml-file name) evalYmod: Function to evaluate yMod for the log-likelhood. evalU0!: Function that computes the initial u0 value for the ODE-system. evalSd!: Function that computes the standard deviation value for the log-likelhood. odeSystem: ModellingToolkit ODE-system for the PeTab model. paramMap: A map to correctly map model parameters to the ODE-system. stateMap: A map to correctly mapping the parameters to the u0 values. paramNames: Names of the model parameters (both fixed and those to be estimated). stateNames: Names of the model states. dirModel: Directory where the model.xml and PeTab files are stored. pathMeasurementData: Path to the measurementData PeTab file. pathMeasurementData: Path to the experimentaCondition PeTab file pathMeasurementData: Path to the observables PeTab file pathMeasurementData: Path to the parameters PeTab file

See also: [setUpCostFunc]

PEtab.SimulationInfo — Type

SimulationInfo

Struct storing simulation (forward ODE-solution) information. Specifcially stores the experimental ID:s from the experimentalCondition - PeTab file; firstExpIds (preequilibration ID:s), the shiftExpIds (postequilibration), and simulateSS (whether or not to simulate ODE-model to steady state). Further stores a solArray with the ODE solution where conditionIdSol of the ID for each forward solution. It also stores for each experimental condition which time-points we have observed data at

See also: [getSimulationInfo]

PEtab.XmlToModellingToolkit — Method

XmlToModellingToolkit(pathXml::String, modelName::String, dirModel::String)

Convert a SBML file in pathXml to a Julia ModelingToolkit file and store
the resulting file in dirModel with name modelName.jl.

The SBML importer goes via libsbml in Python and currently likelly only
works with SBML level 3.

PEtab.calcAccuracyOdeSolver — Method

calcAccuracyOdeSolver(prob::ODEProblem,
                      solArrayHighAccuracy::Vector{<:SciMLBase.AbstractODESolution},
                      changeToExperimentalCondUse!::Function,
                      measurementData::DataFrame,
                      simulationInfo::SimulationInfo,
                      solver,
                      absTol::Float64,
                      relTol::Float64)::Float64

Check the accuracy of an ODE solver at specific tol=absTol=relTol for a PeTab ODE model (odeProb) by for each experimental condition computing the squared sum difference against a high accuracy ODE solution (for each experimental condition) stored in solArrayHighAccuracy.

Recomended to compute high accuracy solution with small tolerances (1e-15) using a high accuracy solver and BigFloat.

PEtab.calcCost — Method

calcCost(paramVecEst,
         odeProb::ODEProblem,
         peTabModel::PeTabModel,
         simulationInfo::SimulationInfo,
         paramIndices::ParameterIndices,
         measurementData::MeasurementData,
         parameterData::ParamData,
         changeModelParamUse!::Function,
         solveOdeModelAllCondUse!::Function;
         calcHessian::Bool=false)

For a PeTab model compute the cost (likelhood) for a parameter vector
paramVecEst. With respect to paramVecEst (all other inputs fixed)
the function is compatible with ForwardDiff.

To compute the cost an ODE-problem, peTabModel, ODE simulation info,
indices to map parameter from paramVecEst, measurement data, parameter
data (e.g constant parameters), function to map parameters correctly to
ODE-model, and a function to solve the ODE model are required. These
are all set up correctly by the `setUpCostGradHess` function.

See also: [`setUpCostGradHess`]

PEtab.calcGradCost! — Method

calcGradCost!(grad::T1,
              paramVecEst,
              odeProb::ODEProblem,
              peTabModel::PeTabModel,
              simulationInfo::SimulationInfo,
              paramIndices::ParameterIndices,
              measurementData::MeasurementData,
              parameterData::ParamData,
              changeModelParamUse!::Function,
              solveOdeModelAllCondUse!::Function) where T1<:Array{<:AbstractFloat, 1}

For a PeTab model compute inplace the gradient  of the cost (likelhood) for
a parameter vector paramVecEst.

Currently the gradient for dynamic parameters (part of ODE-system) is computed via ForwardDiff,
and ReverseDiff is used for observable and sd parameters. The input arguements are the same
as for `calcCost`, and everything is setup by `setUpCostGradHess` function.

See also: [`setUpCostGradHess`]

PEtab.calcHessianApprox! — Method

calcHessianApprox!(hessian::T1,
                   paramVecEst,
                   odeProb::ODEProblem,
                   peTabModel::PeTabModel,
                   simulationInfo::SimulationInfo,
                   paramIndices::ParameterIndices,
                   measurementData::MeasurementData,
                   parameterData::ParamData,
                   changeModelParamUse!::Function,
                   solveOdeModelAllCondUse!::Function) where T1<:Array{<:AbstractFloat, 2}

For a PeTab model compute inplace hessian approximation of the cost (likelhood) for
a parameter vector paramVecEst.

The hessian approximation assumes the interaction betweeen dynamic and (observable, sd) parameters is zero.
The input arguements are the same as for `calcCost`, and everything is setup by `setUpCostGradHess` function.

See also: [`setUpCostGradHess`]

PEtab.calcHighAccOdeSolution — Method

calcHighAccOdeSolution(prob::ODEProblem,
                       changeToExperimentalCondUse!::Function,
                       measurementData::DataFrame,
                       simulationInfo::SimulationInfo;
                       absTol::Float64=1e-15,
                       relTol::Float64=1e-15,
                       nTSave=100)

For a PeTab ODE model with the parameter values in prob compute a high accuracy ODE solution with absTol=relTol=tol using BigFloats for the ode-problem. The ODE is solved for all experimental conditions specified in the PeTab files. Returns an array with ODE-solution and a bool which is true if the ODE model could be solved.

By default the function first tries to solve the ODE problem using a non-stiff solver AutoVern9(Rodas4P()). In case the non-stiff solver fails the Rodas4P() is used to compute the high accuracy solution. If both solvers fail status fail is returned.

PEtab.calcLogLik — Method

calcLogLik(dynamicParamEst::T1,
                sdParamEst,
                obsPar,
                peTabModel::PeTabModel,
                simulationInfo::SimulationInfo,
                paramIndices::ParameterIndices,
                measurementData::MeasurementData,
                parameterData::ParamData;
                gradHessDynParam::Bool=false)::Real where T1<:Vector{<:Real}

Helper function computing the likelhood by given after solving the ODE-system using
using the dynamic-parameters, sd-parameters and observable parameters.

Currently for Gaussian data log10 and non-transformed data is accepted.

See also: [`calcCost`, `setUpCostGradHess`]

PEtab.calcLogLikNotSolveODE — Method

calcLogLikNotSolveODE(dynamicParamEst,
                      sdParamEst,
                      obsParamEst,
                      peTabModel::PeTabModel,
                      simulationInfo::SimulationInfo,
                      paramIndices::ParameterIndices,
                      measurementData::MeasurementData,
                      parameterData::ParamData)

Helper function computing the likelhood by given  an already existing ODE-solution stored
in simulationInfo using the dynamic-parameters, sd-parameters and observable parameters.

When computing the cost and gradient/hessian for the sd- and observable-parameters
only a solved ODE-system is needed (no need to resolve). This greatly reduces run-time
since a lot of dual numbers do not have to be propegated through the ODE solver.
Besides the different parameter vector the input arguements are the same as for `calcCost`,
and everything is setup by `setUpCostGradHess` function.

See also: [`calcCost`, `setUpCostGradHess`]

PEtab.calcLogLikSolveODE — Method

calcLogLikSolveODE(dynamicParamEst,
                   sdParamEst,
                   obsParEst,
                   nonDynParamEst,
                   odeProb::ODEProblem,
                   peTabModel::PeTabModel,
                   simulationInfo::SimulationInfo,
                   measurementData ::MeasurementData,
                   parameterData::ParamData,
                   changeModelParamUse!::Function,
                   solveOdeModelAllCondUse!::Function,
                   computeGradOrHess::Bool=false)

Helper function computing the likelhood by solving the ODE system for all
PeTab-specifed experimental conditions using the dynamic-parameters,
sd-parameters and observable parameters.

When computing the cost and gradient/hessian for dynamic parameters the ODE
system must be solved before getting the likelhood. Besides the different
parameter vector the input arguements are the same as for `calcCost`, and
everything is setup by `setUpCostGradHess` function.

See also: [`calcCost`, `setUpCostGradHess`]

PEtab.calcSqErrVal — Method

calcSqErrVal(solHighAccuracy, sol, t_max::Float64)::Float64

Helper function to compute the squared sum difference betweeen two ODESolutions solved to t_max. Here solHighAccuracy is meant to be a high accuracy solution.

PEtab.changeExperimentalCond! — Method

changeExperimentalCond!(paramVec,
                        stateVec,
                        expID::String,
                        parameterData::ParamData,
                        experimentalConditions::DataFrame,
                        peTabModel::PeTabModel)

Change the ODE parameter vector (paramVec) and initial value vector (stateVec) values to the values specified for the experimental ID expID given by the experimentalConditions peTab-file for a specific peTabModel.

parameterData is needed to correctly map the parameters. The function can handle that paramVec is a Float64 vector or a vector of Duals for the gradient calculations. This function is used by the solveOdeModelAllExperimentalCond!.

See also: [solveOdeModelAllExperimentalCond!]

PEtab.changeModelParam! — Method

changeModelParam!(paramVecOdeModel,
                  stateVecOdeModel,
                  paramVecEst,
                  paramEstNames::Array{String, 1},
                  paramIndices::ParameterIndices,
                  peTabModel::PeTabModel)

Change the ODE parameter vector (paramVecOdeModel) and initial value vector (stateVecOdeModel)
values to the values in parameter vector used for parameter estimation paramVecEst.
Basically, map the parameter-estiamtion vector to the ODE model.

The function can handle that paramVecEst is a Float64 vector or a vector of Duals for the
gradient calculations. This function is used when computing the cost, and everything
is set up by `setUpCostGradHess`.

See also: [`setUpCostGradHess`]

PEtab.checkForPeTabFile — Method

checkForPeTabFile(fileSearchFor::String, dirModel::String)::String

Helper function to check in dirModel if a file starting with fileSearchFor exists.
If true return file path.

PEtab.createBigFloatODEProblem — Method

createBigFloatODEProblem(peTabModel::PeTabModel)::ODEProblem

From a PeTab model create its corresponding ODE-problem with BigFloat (long double).

PEtab.createBigFloatODEProblem — Method

createBigFloatODEProblem(odeProb::ODEProblem)::ODEProblem

Convert an ODE problem with arbitrary float to one an ODE-problem with BigFloat.

PEtab.createFileYmodSdU0 — Method

createFileYmodSdU0(modelName::String,
                   dirModel::String,
                   odeSys::ODESystem,
                   stateMap,
                   modelDict)

For a PeTab model with name modelName with all PeTab-files in dirModel and associated
ModellingToolkit ODESystem (with its stateMap) build a file containing a functions for
i) computing the observable model value (yMod) ii) compute the initial value u0 (by using the
stateMap) and iii) computing the standard error (sd) for each observableFormula in the
observables PeTab file.

Note - The produced Julia file will go via the JIT-compiler.

PEtab.createSdFunction — Method

createSdFunction(modelName::String,
                      dirModel::String,
                      paramData::ParamData,
                      stateNames,
                      namesParamDyn::Array{String, 1},
                      observablesData::DataFrame)

For modelName create a function for computing the standard deviation by translating the observablesData
PeTab-file into Julia syntax.

To correctly create the function the state-names, names of dynamic parameters to estiamte
(namesDynParam) and PeTab parameter-file (to get constant parameters) data are needed.

PEtab.createU0Function — Method

createU0Function(modelName::String,
                     dirModel::String,
                     paramData::ParamData,
                     namesParameter::Array{String, 1},
                     stateMap)

For modelName create a function for computing initial value by translating the stateMap
into Julia syntax.

To correctly create the function the name of all parameters, paramData (to get constant parameters)
are required.

PEtab.createYmodFunction — Method

createYmodFunction(modelName::String,
                   dirModel::String,
                   stateNames,
                   paramData::ParamData,
                   namesParamDyn::Array{String, 1},
                   observablesData::DataFrame,
                   modelDict)

For modelName create a function for computing yMod by translating the observablesData
PeTab-file into Julia syntax.

To correctly create the function the state-names, names of dynamic parameters to estiamte
(namesDynParam) and PeTab parameter-file (to get constant parameters) data are needed.

The modelDict is used to define explicit rules to the Ymod file.

PEtab.getIndicesParam — Method

getIndicesParam(paramData::ParamData, measurementData::MeasurementData)::ParameterIndices

For a PeTab-model creates index vectors for extracting the dynamic-, observable- and sd-parameters from
the  optimizsing parameter vector, and indices and maps to effectively for an observation, when computing
the likelhood, to extract the correct SD- and obs-parameters.

When correctly extracting sd- or obs-parameters the `ParamMap` struct is used. This struct has precomputed
values for which parameters belong to a specific observation, for example, which of the to be estimated
observable parameters should be used when computing a part of the likelihood.

See also: [`ParameterIndices`, `ParamMap`]

PEtab.getNoiseParamStr — Method

getNoiseParamStr(sdFormula::String)::String

Helper function to extract all the noiseParameter in noiseParameter formula in the PeTab file.

PEtab.getObsParamStr — Method

getObsParamStr(measurmentFormula::String)::String

Helper function to extract all observableParameter in the observableFormula in the PeTab-file.

PEtab.getRowExpId — Method

getRowExpId(expId::String, data::DataFrame; colSearch="conditionId")

Small helper function to get which row in a DataFrame corresponds to specific ExpId

PEtab.getSimulationInfo — Method

getSimulationInfo(measurementData::DataFrame)::SimulationInfo

Using the PeTab measurementData-file extract information on the foward ODE simulations.

Specifcially extract the experimental ID:s from the experimentalCondition - PeTab file;
firstExpIds (preequilibration ID:s), the shiftExpIds (postequilibration), and
simulateSS (whether or not to simulate ODE-model to steady state). Further
stores a solArray with the ODE solution where conditionIdSol of the ID for
each forward solution

TODO: Compute t-vec save at from measurementDataFile (instead of providing another struct)

PEtab.getTimeMax — Method

getTimeMax(measurementData::DataFrame, expId::String)::Float64

Small helper function to get the time-max value for a specific simulationConditionId when simulating
the PeTab ODE-model

PEtab.getWord — Method

getWord(str::String, iStart, charListTerm)

In a string starting from position iStart extract the next "word", which is the longest
concurent occurance of characters that are not in the character list with word termination
characters. Returns the word and iEnd (the position where the word ends).

For example, if charListTerm = ['(', ')', '+', '-', '/', '*', '^'] abc123 is
considered a word but not abc123*.

PEtab.isNumber — Method

isNumber(x::AbstractString)::Bool

Check if a string x is a number (Float).

PEtab.peTabFormulaToJulia — Method

peTabFormulaToJulia(formula::String, stateNames, paramData::ParamData, namesParamDyn::Array{String, 1}, namesNonDynParam::Array{String, 1}, namesExplicitRules::Array{String, 1})::String
Translate a peTab formula (e.g for observable or for sd-parameter) into Julia syntax and output the result
as a string.

State-names, namesParamDyn and paramData are all required to correctly identify states and parameters in the formula.
namesExplicitRules is optional and is only set if there are any explicit rules in the SBML-file.

PEtab.processMeasurementData — Method

processMeasurementData(measurementData::DataFrame, observableData::DataFrame)::MeasurementData

Process the PeTab measurementData file into a type-stable Julia struct.

PEtab.processParameterData — Method

processParameterData(parameterData::DataFrame)::ParamData

Process the PeTab parameterData file into a type-stable Julia struct.

PEtab.readDataFiles — Method

readDataFiles(dirModel::String; readObs::Bool=false)

Given a directory for a model, e.g ./Beer_MolBioSystems2014, read the associated PeTab files
for the measurements, parameters, experimental conditions and (if true) the observables.

PEtab.setParamToFileValues! — Method

setParamToFileValues!(paramMap, stateMap, paramData::ParamData)

Function that sets the parameter and state values in paramMap and stateMap
to those in the PeTab parameters file.

Used when setting up the PeTab cost function, and when solving the ODE-system
for the values in the parameters-file.

PEtab.setUpCostGradHess — Method

setUpCostGradHess(peTabModel::PeTabModel, solver, tol::Float64)

For a PeTab-model set up functions for computing i) the likelihood, ii) likelhood gradient,
and iii) likelhood Hessian block approximation. The functions are stored in PeTabOpt-struct
that can be used as input to the optimizers.

Currently the gradient for dynamic parameters (part of ODE-system) is computed via ForwardDiff,
and ReverseDiff is used for observable and sd parameters. The hessian approximation assumes the
interaction betweeen dynamic and (observable, sd) parameters is zero.

PEtab.setUpPeTabModel — Method

setUpPeTabModel(modelName::String, dirModel::String)::PeTabModel

Given a model directory (dirModel) containing the PeTab files and a
xml-file on format modelName.xml will return a PeTabModel struct holding
paths to PeTab files, ode-system in ModellingToolkit format, functions for
evaluating yMod, u0 and standard deviations, and a parameter and state maps
for how parameters and states are mapped in the ModellingToolkit ODE system
along with state and parameter names.

dirModel must contain a SBML file named modelName.xml, and files starting with
measurementData, experimentalCondition, parameter, and observables (tsv-files).
The latter files must be unique (e.g only one file starting with measurementData)

TODO : Example

PEtab.solveOdeModelAllExperimentalCond! — Method

solveOdeModelAllExperimentalCond!(solArray::Vector{<:SciMLBase.AbstractODESolution},
                                  prob::ODEProblem,
                                  changeToExperimentalCondUse!::Function,
                                  measurementData::DataFrame,
                                  simulationInfo::SimulationInfo,
                                  solver,
                                  tol::Float64;
                                  nTSave::Int64=0,
                                  denseSol::Bool=true)::Bool

Solve a PeTab ODE model for all experimental conditions specified in the PeTab experimentaCondition-file
and store each ODE-soluation into solArray. Returns true if model could be solved succesfully for all
conditions, else returns false.

All inputs are automatically computed by the PeTab importer. Each experimental condition is simulated to the
maximum time for that condition found in the the measurementData PeTab-file. In case a pre-equlibration condition
exists the model is first simulated to a steady-state. Then starting from the steady state the solution stored in
solArray is calculcated.

# Args
`solArray`: array storing ODE-solution for each experimental condition. Is pre-allocated by PeTab importer.
`prob`: ODEProblem struct for the model to be simulated.
`changeToExperimentalCondUse!`: function that changes the parameters for the ODE-problem to those for a specific experimental condition
`measurementData`: the PeTab measurementData file
`simulationInfo`: struct storing simulation info like experimental conditions, whether or not to simulate to a steady state.
`solver`: ode-solver for solving the ode-problem. Any Julia solver or alg-hint works.
`tol`: rel- and abs-tol for the ODE solver
`nTSave`: number of equidistant data-points to save for each ODE-solution. If 0 the ODE-solver outputs a dense solution.
`denseSol`: whether to have a dense (true) or none-dense (false) ODE-solution for each experimental condition.

See also: [`setUpCostFunc`, `SimulationInfo`, `changeToExperimentalCond!`]

PEtab.solveOdeModelAtFileValues — Method

solveOdeModelAtFileValues(peTabModel::PeTabModel)

For a peTab model solve the ODE:s at the values in the parameters-PeTab-file
using a specific solver with absTol=relTol=tol. Returns an array with the
solution for each experimenta condition along with a vector with the condition
name for each solution

PEtab.solveOdeNoSS — Method

solveOdeNoSS(prob::ODEProblem,
             changeToExperimentalCondUse!::Function,
             firstExpId::String,
             tol::Float64,
             t_max::Float64,
             solver;
             tSave=Float64[],
             nTSave=0,
             denseSol=true)

For an experimentaCondition specifed by firstExpId solve a PeTab ODE-problem using any Julia ODE-solver
(solver can also be an alg-hint) using absTol and relTol value given by tol. Returns an ODE-solution.

Here the model is not simulated to steady state first first simualted to a steady state
Optional args are the save as for `solveOdeModelAllExperimentalCond!`, and this function
is called by `solveOdeModelAllExperimentalCond!`.

See also: [`solveOdeModelAllExperimentalCond!`]

PEtab.solveOdeSS — Method

solveOdeSS(prob::ODEProblem,
           changeToExperimentalCondUse!::Function,
           firstExpId::String,
           shiftExpId::String,
           tol::Float64,
           t_max_ss::Float64,
           solver;
           tSave=Float64[],
           nTSave=0,
           denseSol=true)

For an experimentaCondition specifed by firstExpId (preequilibration ID) and shiftExpId
(postequilibration ID) solve a PeTab ODE-problem using any Julia ODE-solver (solver can
also be an alg-hint) using absTol and relTol value given by tol. Returns an ODE-solution.

Here the model is first simualted to a steady state using the experimental-condition parameters
corresponding to firstExpId in the experimentaCondition peTab-file. Then using the experimental-condition
parameters specifed by shiftExpId the actual simulation that is saved and returned is produced.
Optional args are the save as for `solveOdeModelAllExperimentalCond!`, and this function is called
by solveOdeModelAllExperimentalCond!.

See also: [`solveOdeModelAllExperimentalCond!`]

PEtab.transformObsOrData — Method

transformObsOrData(val, transform::Symbol)

Transform val using either :lin (identity), :log10 and :log transforamtions.

PEtab.transformParamVec! — Method

transformParamVec!(paramVec, namesParam::Array{String, 1}, paramData::ParamData; revTransform::Bool=false)

Helper function which transforms in-place a parameter vector with parameters specied in namesParam according to the
transformation for said parameter specifid in paramData.shouldTransform. In case revTransform is true
performs the inverse parameter transformation (e.g exp10 instead of log10)

PEtab.transformParamVec — Method

transformParamVec!(paramVec, namesParam::Array{String, 1}, paramData::ParamData; revTransform::Bool=false)

Helper function which returns a transformed parameter vector with parameters specied in namesParam according to the
transformation for said parameter specifid in paramData.shouldTransform. In case revTransform is true
performs the inverse parameter transformation (e.g exp10 instead of log10).

The function is fully compatible with Zygote.

PEtab.transformYobsOrYmodArr! — Method

transformYobsOrYmodArr!(vals, transformationArr::Vector{Symbol})

Transform the Yobs or Ymod arrays (vals) in place using for each value in vals the transformation specifed in transformationArr.

Currently :lin, :log10 and :log transforamtions are supported, see setUpCostFunc.

PEtab.wordToJuliaSyntax — Method

" wordToJuliaSyntax(wordTranslate::String, stateNames, paramData::ParamData, namesParamDyn::Array{String, 1}, namesExplicitRules::Array{String, 1})::String

Translate a word (state, parameter, math-expression or number) into Julia syntax
when building Ymod, U0 and Sd functions.
namesExplicitRules is optional and is only set if there are any explicit rules in the SBML-file.

PEtab.writeODEModelToFile — Method

writeODEModelToFile(modelDict, modelName, dirModel)

Takes a modelDict as defined by buildODEModelDictionary
and creates a Julia ModelingToolkit file and stores
the resulting file in dirModel with name modelName.jl.