PEtab

Documentation for PEtab.

PEtab.MeasurementsInfo

PEtab.PEtabModel

PEtab.ParameterIndices

PEtab.ParametersInfo

PEtab.JLToModellingToolkit

PEtab.XmlToModellingToolkit

PEtab.checkDataFrameColumns

PEtab.checkForPeTabFile

PEtab.createDerivative_σ_h_File

PEtab.createTopOfFunction_h

PEtab.createTopOf∂h∂_Function

PEtab.create_h_Function

PEtab.create_u0_Function

PEtab.create_σ_Function

PEtab.create_σ_h_u0_File

PEtab.create∂h∂_Function

PEtab.create∂σ∂_Function

PEtab.dualToFloat

PEtab.dualToFloat

PEtab.getNoiseParametersStr

PEtab.getObservableParametersStr

PEtab.getWord

PEtab.isNumber

PEtab.isNumber

PEtab.petabFormulaToJulia

PEtab.processMeasurements

PEtab.processParameters

PEtab.readPEtabModel

PEtab.replaceExplicitVariableWithRule

PEtab.replaceVariablesWithArrayIndex

PEtab.replaceWholeWordWithNumberPrefix

PEtab.setParamToFileValues!

PEtab.transformMeasurementOrH

PEtab.wordToJuliaSyntax

PEtab.writeODEModelToFile

PEtab.MeasurementsInfo — Type

MeasurementData

Struct storing the data in the PeTab measurementData-file in type-stable manner.

See also: [`processMeasurementData`]

PEtab.PEtabModel — Type

PEtabModel

Struct storing information about a PeTab-model. Create by the `setUpCostFunc` function.

# Args
`modelName`: PeTab model name (must match the xml-file name)
`evalYmod`: Function to evaluate yMod for the log-likelhood.
`evalU0!`: Function that computes the initial u0 value for the ODE-system.
`evalSd!`: Function that computes the standard deviation value for the log-likelhood.
`odeSystem`: ModellingToolkit ODE-system for the PeTab model.
`paramMap`: A map to correctly map model parameters to the ODE-system.
`stateMap`: A map to correctly mapping the parameters to the u0 values.
`paramNames`: Names of the model parameters (both fixed and those to be estimated).
`stateNames`: Names of the model states.
`dirModel`: Directory where the model.xml and PeTab files are stored.
`pathMeasurementData`: Path to the measurementData PeTab file.
`pathMeasurementData`: Path to the experimentaCondition PeTab file
`pathMeasurementData`: Path to the observables PeTab file
`pathMeasurementData`: Path to the parameters PeTab file

See also: [`setUpCostFunc`]

PEtab.ParameterIndices — Type

ParameterIndices

Struct storing names and mapping indices for mapping the parameter provided
to the optimizers correctly.

Optimizers require a single vector input of parameters (pVecEst). However, the PeTab
model has three kind of parameters, Dynmaic (part of the ODE-system),
Observable (only part of the observation model) and Standard-deviation
(only part of the standard deviation expression in the log-likelhood). This
struct stores mapping indices (starting with i) to map pVecEst
correctly when computing the likelihood (e.g map the SD-parameters in pVecEst
correctly to a vector of SD-vals). Also stores the name of each parameter.

Furthermore, when computing yMod or SD the correct observable and sd parameters
has to be used for each observation. The mapArrays effectively contains precomputed
maps allowing said parameter to be effectively be extracted by the getObsOrSdParam
function.

See also: [`getIndicesParam`, `ParamMap`]

PEtab.ParametersInfo — Type

ParameterInfo

Struct storing the data in the PeTab parameter-file in type-stable manner.

Currently logScale notices whether or not parameters are estimated on the
log10 scale or not.

See also: [`processParameterData`]

PEtab.JLToModellingToolkit — Method

JLToModellingToolkit(modelName::String, dirModel::String)

Checks and fixes a Julia ModelingToolkit file and store
the fixed file in dirModel with name modelName_fix.jl.

PEtab.XmlToModellingToolkit — Method

XmlToModellingToolkit(pathXml::String, modelName::String, dirModel::String)

Convert a SBML file in pathXml to a Julia ModelingToolkit file and store
the resulting file in dirModel with name modelName.jl.

The SBML importer goes via libsbml in Python and currently likelly only
works with SBML level 3.

PEtab.checkDataFrameColumns — Method

checkDataFrameColumns(dataFrame, dataFrameName, colsToCheck, allowedTypesVec, requiredCols) Goes through each column from colsToCheck in dataFrame and checks if each column is of any of the DataTypes specified in allowedTypesVec[colIndex]. Returns true if all columns are ok and false otherwise. requiredCols is an array of mandatory columns. If a mandatory column is missing an error is thrown, and if a mandatory column contains missing rows a warning is thrown.

PEtab.checkForPeTabFile — Method

checkForPeTabFile(fileSearchFor::String, dirModel::String)::String

Helper function to check in dirModel if a file starting with fileSearchFor exists. If true return file path.

PEtab.createDerivative_σ_h_File — Method

createFileDYmodSdU0(modelName::String,
                   dirModel::String,
                   odeSys::ODESystem,
                   stateMap,
                   SBMLDict::Dict)

For a PeTab model with name modelName with all PeTab-files in dirModel and associated
ModellingToolkit ODESystem (with its stateMap) build a file containing a functions for
i) computing the observable model value (yMod) ii) compute the initial value u0 (by using the
stateMap) and iii) computing the standard error (sd) for each observableFormula in the
observables PeTab file.
Note - The produced Julia file will go via the JIT-compiler.

PEtab.createTopOfFunction_h — Method

createTopOfFunction_h(modelStateNames::Vector{String},
                      paramData::ParametersInfo,
                      namesParamODEProb::Vector{String},
                      namesNonDynParam::Vector{String})

Extracts all variables needed for the observable h function.

PEtab.createTopOf∂h∂_Function — Method

createTopOf∂h∂_Function(modelStateNames::Vector{String},
                        pODEProblemNames::Vector{String},
                        θ_nonDynamicNames::Vector{String},
                        observablesData::DataFrame)

Extracts all variables needed for the functions and add them as variables for Symbolics.

PEtab.create_h_Function — Method

create_h_Function(modelName::String,
                  dirModel::String,
                  modelStateNames::Vector{String},
                  paramData::ParametersInfo,
                  namesParamDyn::Vector{String},
                  namesNonDynParam::Vector{String},
                  observablesData::DataFrame,
                  SBMLDict::Dict)

For modelName create a function for computing yMod by translating the observablesData
PeTab-file into Julia syntax.

PEtab.create_u0_Function — Method

create_u0_Function(modelName::String,
                   dirModel::String,
                   parameterInfo::ParametersInfo,
                   pODEProblemNames::Vector{String},
                   stateMap;
                   inPlace::Bool=true)

For modelName create a function for computing initial value by translating the stateMap
into Julia syntax.

To correctly create the function the name of all parameters, paramData (to get constant parameters)
are required.

PEtab.create_σ_Function — Method

create_σ_Function(modelName::String,
                  dirModel::String,
                  parameterInfo::ParametersInfo,
                  modelStateNames::Vector{String},
                  pODEProblemNames::Vector{String},
                  θ_nonDynamicNames::Vector{String},
                  observablesData::DataFrame,
                  SBMLDict::Dict)

For modelName create a function for computing the standard deviation σ by translating the observablesData
PeTab-file into Julia syntax.

PEtab.create_σ_h_u0_File — Method

create_σ_h_u0_File(modelName::String,
                   dirModel::String,
                   odeSystem::ODESystem,
                   stateMap,
                   SBMLDict::Dict;
                   verbose::Bool=false)

For a PeTab model with name modelName with all PeTab-files in dirModel and associated
ModellingToolkit ODESystem (with its stateMap) build a file containing a functions for
i) computing the observable model value (h) ii) compute the initial value u0 (by using the
stateMap) and iii) computing the standard error (σ) for each observableFormula in the
observables PeTab file.

Note - The produced Julia file will go via the JIT-compiler. The SBML-dict is needed as
sometimes variables are encoded via explicit-SBML rules.

PEtab.create∂h∂_Function — Method

create∂h∂_Function(modelName::String,
                   dirModel::String,
                   modelStateNames::Vector{String},
                   parameterInfo::ParametersInfo,
                   pODEProblemNames::Vector{String},
                   θ_nonDynamicNames::Vector{String},
                   observablesData::DataFrame,
                   SBMLDict::Dict)

For modelName create using Symbolics function for computing ∂h/∂u and ∂h/∂p where
u = modelStates and p = pODEProblem (parameters for ODE problem)

PEtab.create∂σ∂_Function — Method

create∂σ∂_Function(modelName::String,
                        dirModel::String,
                        parameterInfo::ParametersInfo,
                        modelStateNames::Vector{String},
                        pODEProblemNames::Vector{String},
                        θ_nonDynamicNames::Vector{String},
                        observablesData::DataFrame,
                        SBMLDict::Dict)

For modelName create a function for computing the standard deviation by translating the observablesData

PEtab.dualToFloat — Method

dualToFloat(x::AbstractFloat)::AbstractFloat

PEtab.dualToFloat — Method

dualToFloat(x::ForwardDiff.Dual)::Real

Via recursion convert a Dual to a Float.

PEtab.getNoiseParametersStr — Method

getNoiseParametersStr(formula::String)::String

Helper function to extract all the noiseParameter in noiseParameter formula in the PEtab file.

PEtab.getObservableParametersStr — Method

getObservableParametersStr(formula::String)::String

Helper function to extract all observableParameter in the observableFormula in the PEtab-file.

PEtab.getWord — Method

getWord(str::String, iStart::Int, charTerminate::Vector{Char})

In a string starting from position iStart extract the next "word", which is the longest
concurent occurance of characters that are not in the character list with word termination
characters. Returns the word and iEnd (the position where the word ends).

For example, if charListTerm = ['(', ')', '+', '-', '/', '*', '^'] abc123 is
considered a word but not abc123*.

PEtab.isNumber — Method

isNumber(x::String)::Bool

Check if a string x is a number (Float) taking sciencetific notation into account.

PEtab.isNumber — Method

isNumber(x::SubString{String})::Bool

PEtab.petabFormulaToJulia — Method

petabFormulaToJulia(formula::String, stateNames, paramData::ParametersInfo, namesParamDyn::Vector{String}, namesNonDynParam::Vector{String})::String

Translate a peTab formula (e.g for observable or for sd-parameter) into Julia syntax and output the result
as a string.

PEtab.processMeasurements — Method

processMeasurements(measurementsFile::DataFrame, observableData::DataFrame)::MeasurementData

Process the PeTab measurementsFile file into a type-stable Julia struct.

PEtab.processParameters — Method

processParameters(parametersFile::DataFrame)::ParameterInfo

Process the PeTab parametersFile file into a type-stable Julia struct.

PEtab.readPEtabModel — Method

setUpPeTabModel(modelName::String, dirModel::String)::PEtabModel
Given a model directory (dirModel) containing the PeTab files and a
xml-file on format modelName.xml will return a PEtabModel struct holding
paths to PeTab files, ode-system in ModellingToolkit format, functions for
evaluating yMod, u0 and standard deviations, and a parameter and state maps
for how parameters and states are mapped in the ModellingToolkit ODE system
along with state and parameter names.
dirModel must contain a SBML file named modelName.xml, and files starting with
measurementData, experimentalCondition, parameter, and observables (tsv-files).
The latter files must be unique (e.g only one file starting with measurementData)
TODO : Example

PEtab.replaceExplicitVariableWithRule — Method

replaceExplicitVariableWithRule(formula::String, SBMLDict::Dict)::String

Replace the explicit rule variable with the explicit rule

PEtab.replaceVariablesWithArrayIndex — Method

replaceVariablesWithArrayIndex(formula,stateNames,parameterNames,namesNonDynParam,paramData)::String

Replaces any state or parameter from formula with their corresponding index in the ODE system
Symbolics can return strings without multiplication sign, e.g. 100.0STAT5 instead of 100.0*STAT5
so replaceWholeWord cannot be used here

PEtab.replaceWholeWordWithNumberPrefix — Method

replaceWholeWordWithNumberPrefix(formula, from, to)::String

Replaces variables that can be prefixed with numbers, e.g.,
replaceWholeWordWithNumberPrefix("4STAT5 + 100.0STAT5 + RE*STAT5 + STAT5","STAT5","u[1]") gives
4u[1] + 100.0u[1] + RE*u[1] + u[1]

PEtab.setParamToFileValues! — Method

setParamToFileValues!(paramMap, stateMap, paramData::ParamData)

Function that sets the parameter and state values in paramMap and stateMap
to those in the PeTab parameters file.

Used when setting up the PeTab cost function, and when solving the ODE-system
for the values in the parameters-file.

PEtab.transformMeasurementOrH — Method

transformMeasurementOrH(val::Real, transformationArr::Array{Symbol, 1})

Transform val using either :lin (identify), :log10 and :log transforamtions.

PEtab.wordToJuliaSyntax — Method

" wordToJuliaSyntax(wordTranslate::String, stateNames, paramData::ParametersInfo, namesParamDyn::Vector{String})::String

Translate a word (state, parameter, math-expression or number) into Julia syntax
when building Ymod, U0 and Sd functions.

PEtab.writeODEModelToFile — Method

writeODEModelToFile(modelDict, modelName, dirModel)

Takes a modelDict as defined by buildODEModelDictionary
and creates a Julia ModelingToolkit file and stores
the resulting file in dirModel with name modelName.jl.