API

PEtab.PEtabModel — Type

PEtabModel

A Julia-compatible representation of a PEtab-specified problem.

For how to construct see below.

Note

Most of the functions in PEtabModel are not intended to be accessed by the user. For example, compute_h (and similar functions) require indices that are built in the background to efficiently map parameters between experimental (simulation) conditions. Rather, PEtabModel holds all information needed to create a PEtabODEProblem, and in the future, PEtabSDEProblem, etc.

Fields

model_name: The model name extracted from the PEtab YAML file.

compute_h: Computes the observable h for a specific time point and simulation condition.

compute_u0!: Computes in-place initial values using ODEProblem.p for a simulation condition; compute_u0!(u0, p).

compute_u0: Computes initial values as above, but not in-place; u0 = compute_u0(p).

compute_σ: Computes the noise parameter σ for a specific time point and simulation condition.

compute_∂h∂u!: Computes the gradient of h with respect to ODEModel states (u) for a specific time point and simulation condition.

compute_∂σ∂u!: Computes the gradient of σ with respect to ODEModel states (u) for a specific time point and simulation condition.

compute_∂h∂p!: Computes the gradient of h with respect to ODEProblem.p.

compute_∂σ∂p!: Computes the gradient of σ with respect to ODEProblem.p.

compute_tstops: Computes the event times in case the model has DiscreteCallbacks (events).

convert_tspan::Bool: Tracks whether the time span should be converted to Dual numbers for ForwardDiff.jl gradients, in case the model has DiscreteCallbacks and the trigger time is a parameter set to be estimated.

dir_model: The directory where the model.xml and PEtab files are stored.

dir_julia: The directory where the Julia-model files created by parsing the PEtab files (e.g., SBML file) are stored.

ode_system: A ModellingToolkit.jl ODE system obtained from parsing the model SBML file.

parameter_map: A ModellingToolkit.jl parameter map for the ODE system.

state_map: A ModellingToolkit.jl state map for the ODE system describing how the initial values are computed, e.g., whether or not certain initial values are computed from parameters in the parameter_map.

parameter_names: The names of the parameters in the ode_system.

state_names: The names of the states in the ode_system.

path_measurements: The path to the PEtab measurements file.

path_conditions: The path to the PEtab conditions file.

path_observables: The path to the PEtab observables file.

path_parameters: The path to the PEtab parameters file.

path_SBML: The path to the PEtab SBML file.

path_yaml: The path to the PEtab YAML file.

model_callbacks: This stores potential model callbacks or events.

check_callback_is_active: Piecewise SBML statements are transformed to DiscreteCallbacks that are activated at a specific time-point. The piecewise callback has a default value at t0 and is only triggered when reaching tactivation. If tactivation ≤ 0 (never reached when solving the model), this function checks whether the callback should be triggered before solving the model.

Missing docstring.

Missing docstring for PEtabModel. Check Documenter's build log for details.

PEtab.PEtabODEProblem — Type

Everything needed to setup an optimization problem (compute cost, gradient, hessian and parameter bounds) for a PEtab model.

For how to construct, see below.

Note

The parameter vector θ is always assumed to be on the parameter scale specified in the PEtab parameters file. If needed, θ is transformed to the linear scale inside the function call.

Fields

compute_cost: For θ computes the negative likelihood (objective to minimize)

compute_chi2: For θ compute χ2 value

compute_gradient!: For θ computes in-place gradient compute_gradient!(gradient, θ)

compute_gradient: For θ computes out-place gradient gradient = compute_gradient(θ)

compute_hessian!: For θ computes in-place hessian-(approximation) compute_hessian!(hessian, θ)

compute_hessian: For θ computes out-place hessian-(approximation) hessian = compute_hessian(θ)

compute_simulated_values: For θ compute the corresponding model (simulated) values to the measurements in the same order as in the Measurements PEtab table

compute_residuals: For θ compute the residuals (hmodel - hobserved)^2 / σ^2 in the same order as in the Measurements PEtab table

gradient_method: The method used to compute the gradient (either :ForwardDiff, :ForwardEquations, :Adjoint, or :Zygote).

hessian_method: The method used to compute or approximate the Hessian (either :ForwardDiff, :BlocForwardDiff, or :GaussNewton).

n_parameters_esimtate: The number of parameters to estimate.

θ_names: The names of the parameters in θ.

θ_nominal: The nominal values of θ as specified in the PEtab parameters file.

θ_nominalT: The nominal values of θ on the parameter scale (e.g., log) as specified in the PEtab parameters file.

lower_bounds: The lower parameter bounds on the parameter scale for θ as specified in the PEtab parameters file.

upper_bounds: The upper parameter bounds on the parameter scale for θ as specified in the PEtab parameters file.

petab_model: The PEtabModel used to construct the PEtabODEProblem.

ode_solver: The options for the ODE solver specified when creating the PEtabODEProblem.

ode_solver_gradient: The options for the ODE solver gradient specified when creating the PEtabODEProblem.

Missing docstring.

Missing docstring for PEtabODEProblem. Check Documenter's build log for details.

PEtab.PEtabObservable — Type

PEtabObservable(obs_formula, noise_formula; transformation::Symbol=:lin)

Links a model to measurements using an observable formula and measurement noise formula.

The transformation argument can take one of three values: :lin (for normal measurement noise), :log, or :log10 (for log-normal measurement noise). For a full description of options, including how to define measurement-specific observable and noise parameters, see the main documentation.

Examples

# Example 1: Log-normal measurement noise with known error σ=3.0
@unpack X = rn  # 'rn' is the dynamic model
PEtabObservable(X, 3.0, transformation=:log)

# Example 2: Normal measurement noise with estimation of σ (defined as PEtabParameter)
@unpack X, Y = rn  # 'rn' is the dynamic model
@parameters sigma
PEtabObservable((X + Y) / X, sigma)

# Example 3: Normal measurement noise with measurement-specific noiseParameter
@unpack X, Y = rn  # 'rn' is the dynamic model
@parameters noiseParameter1  # Must be in the format 'noiseParameter'
PEtabObservable(X, noiseParameter1 * X)

PEtab.PEtabParameter — Type

PEtabParameter(id::Symbol; <keyword arguments>)

Represents a parameter to be estimated in a PEtab model calibration problem.

Keyword Arguments

estimate::Bool=true: Specifies whether the parameter should be estimated (default) or set as constant.

value::Union{Nothing, Float64}=nothing: The parameter value to use if estimate=false. Defaults to the midpoint between lb and ub.

scale::Symbol=:log10: The scale on which to estimate the parameter. Allowed options are :log10 (default), :log, and :lin.

lb::Float64=1e-3: The lower parameter bound in parameter estimation (default: 1e-3).

ub::Float64=1e-3: The upper parameter bound in parameter estimation (default: 1e3).

prior=nothing: An optional continuous prior distribution from the Distributions package.

prior_on_linear_scale::Bool=true: Specifies whether the prior is on the linear scale (default) or the transformed scale, e.g., log10-scale.

Examples

# Example 1: Parameter with a Log-Normal prior (LN(μ=3.0, σ=1.0)) estimated on the log10 scale
PEtabParameter(:c1, prior=LogNormal(3.0, 1.0))

# Example 2: Parameter estimated on the log scale with a Normal prior (N(0.0, 1.0)) on the log scale
PEtabParameter(:c1, scale=:log, prior=Normal(0.0, 1.0), prior_on_linear_scale=false)

PEtab.ODESolver — Type

ODESolver(solver, <keyword arguments>)

ODE-solver options (solver, tolerances, etc...) to use when computing gradient/cost for a PEtabODEProblem.

More information about the available options and solvers can be found in the documentation for DifferentialEquations.jl (https://docs.sciml.ai/DiffEqDocs/stable/solvers/ode_solve/). Recommended settings for which solver and options to use for different problems can be found below and in the documentation.

Arguments

solver: Any of the ODE solvers in DifferentialEquations.jl. For small (≤20 states) mildly stiff models, composite solvers such as AutoVern7(Rodas5P()) perform well. For stiff small models, Rodas5P() performs well. For medium-sized models (≤75 states), QNDF(), FBDF(), and CVODE_BDF() perform well. CVODE_BDF() is not compatible with automatic differentiation and thus cannot be used if the gradient is computed via automatic differentiation or if the Gauss-Newton Hessian approximation is used. If the gradient is computed via adjoint sensitivity analysis, CVODE_BDF() is often the best choice as it is typically more reliable than QNDF() and FBDF() (fails less often).

abstol=1e-8: Absolute tolerance when solving the ODE system. Not recommended to increase above 1e-6 for gradients.

reltol=1e-8: Relative tolerance when solving the ODE system. Not recommended to increase above 1e-6 for gradients.

force_dtmin=false: Whether or not to force dtmin when solving the ODE system.

dtmin=nothing: Minimal acceptable step-size when solving the ODE system.

maxiters=10000: Maximum number of iterations when solving the ODE system. Increasing above the default value can cause the optimization to take substantial time.

PEtab.SteadyStateSolver — Type

SteadyStateSolver(method::Symbol;
                  check_simulation_steady_state::Symbol=:wrms,
                  rootfinding_alg=nothing,
                  abstol=nothing,
                  reltol=nothing,
                  maxiters=nothing)

Setup options for finding steady-state via either method=:Rootfinding or method=:Simulate.

For method=:Rootfinding, the steady-state u* is found by solving the problem du = f(u, p, t) ≈ 0 with tolerances abstol and reltol via an automatically chosen optimization algorithm (rootfinding_alg=nothing) or via any provided algorithm in NonlinearSolve.jl.

For method=:Simulate, the steady-state u* is found by simulating the ODE system until du = f(u, p, t) ≈ 0. Two options are available for check_simulation_steady_state:

:wrms : Weighted root-mean square √(∑((du ./ (reltol * u .+ abstol)).^2) / length(u)) < 1

:Newton : If Newton-step Δu is sufficiently small √(∑((Δu ./ (reltol * u .+ abstol)).^2) / length(u)) < 1. - Newton often performs better but requires an invertible Jacobian. In case it's not fulfilled, the code switches automatically to :wrms.

maxiters refers to either the maximum number of rootfinding steps or the maximum number of integration steps, depending on the chosen method.

PEtab.remake_PEtab_problem — Function

remake_PEtab_problem(petab_problem::PEtabODEProblem, parameters_change::Dict) :: PEtabODEProblem

Fixate model parameters for a given PEtabODEProblem without recompiling the problem.

This function allows you to modify parameters without the need to recompile the underlying code, resulting in reduced latency. To fixate the parameter k1, you can use parameters_change=Dict(:k1 => 1.0).

If model derivatives are computed using ForwardDiff.jl with a chunk-size of N, the new PEtabODEProblem will only evaluate the necessary number of chunks of size N to compute the full gradient for the remade problem.

PEtab.Fides — Type

Fides(hessian_method::Union{Nothing, Symbol}; verbose::Bool=false)

Fides is a Python Newton-trust region optimizer designed for box-bounded optimization problems.

It excels when the full Hessian is too computationally expensive, but a Gauss-Newton Hessian approximation can be calculated. When constructed with Fides(verbose=true), it displays optimization progress during estimation.

Hessian Methods

If hessian_method=nothing, the Hessian method from the PEtabODEProblem is used, which can be either exact or a Gauss-Newton approximation. Additionally, Fides supports the following Hessian approximation methods:

:BB: Broyden's "bad" method

:BFGS: Broyden-Fletcher-Goldfarb-Shanno update strategy

:BG: Broyden's "good" method

:Broyden: BroydenClass Update scheme

:SR1: Symmetric Rank 1 update

:SSM: Structured Secant Method

:TSSM: Totally Structured Secant Method

For more information on each method, see the Fides documentation.

See also callibrateModel and callibrateModelMultistart

Examples

# Fides with the Hessian method as in the PEtabProblem
fides_opt = Fides(nothing)

# Fides with the BFGS Hessian approximation, with progress printing
fides_opt = Fides(:BFGS; verbose=true)

PEtab.IpoptOptions — Type

IpoptOptions(;print_level::Int64=0, 
             max_iter::Int64=1000, 
             tol::Float64=1e-8, 
             acceptable_tol::Float64=1e-6, 
             max_wall_time::Float64=1e20, 
             acceptable_obj_change_tol::Float64=1e20)

Wrapper for a subset of Ipopt options to set during parameter estimation.

For more information about each options see the Ipopt documentation

Arguments

print_level: Output verbosity level (valid values are 0 ≤ print_level ≤ 12)

max_iter: Maximum number of iterations

tol: Relative convergence tolerance

acceptable_tol: Acceptable relative convergence tolerance

max_wall_time: Max wall time optimisation is allowed to run

acceptable_obj_change_tol: Acceptance stopping criterion based on objective function change.

See also calibrate_model and calibrate_model_multistart.

PEtab.IpoptOptimiser — Type

IpoptOptimiser(LBFGS::Bool)

Ipopt is an Interior-point Newton method designed for nonlinear optimization.

Ipopt can be configured to use either the Hessian method from the PEtabODEProblem (LBFGS=false) or a LBFGS scheme (LBFGS=true). For setting Ipopt options, see IpoptOptions.

See also calibrate_model and calibrate_model_multistart.

Examples

# Ipopt with the Hessian method as in the PEtabProblem
ipopt_opt = IpoptOptimiser(false)

# Ipopt with LBFGS Hessian approximation
ipopt_opt = IpoptOptimiser(true)

PEtab.calibrate_model — Function

calibrate_model(petab_problem::PEtabODEProblem,
                p0::Vector{Float64},
                alg;
                save_trace::Bool=false,
                options=algOptions)::PEtabOptimisationResult

Parameter estimate a model for a PEtabODEProblem using an optimization algorithm alg and an initial guess p0.

The optimization algorithm alg can be one of the following:

Optim LBFGS, BFGS, or IPNewton methods

IpoptOptimiser interior-point optimizer

Fides Newton trust region method

Each algorithm accepts specific optimizer options in the format of the respective package. For a comprehensive list of available options, please refer to the main documentation.

If you want the optimizer to return parameter and objective trace information, set save_trace=true. Results are returned as a PEtabOptimisationResult, which includes the following information: minimum parameter values found (xmin), smallest objective value (fmin), number of iterations, runtime, whether the optimizer converged, and optionally, the trace.

Note

To use Optim optimizers, you must load Optim with using Optim. To use Ipopt, you must load Ipopt with using Ipopt. To use Fides, load PyCall with using PyCall and ensure Fides is installed (see documentation for setup).

Examples

# Perform parameter estimation using Optim's IPNewton with a given initial guess
using Optim
res = calibrate_model(petab_problem, p0, Optim.IPNewton();
                     options=Optim.Options(iterations = 1000))

# Perform parameter estimation using Fides with a given initial guess
using PyCall
res = calibrate_model(petab_problem, p0, Fides(nothing);
                     options=py"{'maxiter' : 1000}"o)

# Perform parameter estimation using Ipopt and save the trace
using Ipopt
res = calibrate_model(petab_problem, p0, IpoptOptimiser(false);
                     options=IpoptOptions(max_iter = 1000), 
                     save_trace=true)

PEtab.calibrate_model_multistart — Function

calibrate_model_multistart(petab_problem::PEtabODEProblem,
                         alg,
                         n_multistarts::Signed,
                         dir_save::Union{Nothing, String};
                         sampling_method=QuasiMonteCarlo.LatinHypercubeSample(),
                         options=algOptions,
                         seed=nothing,
                         save_trace::Bool=false)::PEtabMultistartOptimisationResult

Perform multistart optimization for a PEtabODEProblem using the algorithm alg.

The optimization algorithm alg can be one of the following:

Optim LBFGS, BFGS, or IPNewton methods

IpoptOptimiser interior-point optimizer

Fides Newton trust region method

For each algorithm, optimizer options can be provided in the format of the respective package. For a comprehensive list of available options, please refer to the main documentation. If you want the optimizer to return parameter and objective trace information, set save_trace=true.

Multistart optimization involves generating multiple starting points for optimization runs. These starting points are generated using the specified sampling_method from QuasiMonteCarlo.jl, with the default being LatinHypercubeSample, a method that typically produces better results than random sampling. For reproducibility, you can set a random number generator seed using the seed parameter.

If dir_save is provided as nothing, results are not written to disk. Otherwise, if a directory path is provided, results are written to disk. Writing results to disk is recommended in case the optimization process is terminated after a number of optimization runs.

The results are returned as a PEtabMultistartOptimisationResult, which stores the best-found minima (xmin), smallest objective value (fmin), as well as optimization results for each run.

Note

To use Optim optimizers, you must load Optim with using Optim. To use Ipopt, you must load Ipopt with using Ipopt. To use Fides, load PyCall with using PyCall and ensure Fides is installed (see documentation for setup).

Examples

# Perform 100 optimization runs using Optim's IPNewton, save results in dir_save
using Optim
dir_save = joinpath(@__DIR__, "Results")
res = calibrate_model_multistart(petab_problem, Optim.IPNewton(), 100, dir_save;
                               options=Optim.Options(iterations = 1000))

# Perform 100 optimization runs using Fides, save results in dir_save
using PyCall
dir_save = joinpath(@__DIR__, "Results")
res = calibrate_model_multistart(petab_problem, Fides(nothing), 100, dir_save;
                               options=py"{'maxiter' : 1000}"o)

# Perform 100 optimization runs using Ipopt, save results in dir_save. For each 
# run save the trace 
using Ipopt
dir_save = joinpath(@__DIR__, "Results")
res = calibrate_model_multistart(petab_problem, IpoptOptimiser(false), 100, dir_save;
                               options=IpoptOptions(max_iter = 1000), 
                               save_trace=true)

PEtab.run_PEtab_select — Function

run_PEtab_select(path_yaml, alg; <keyword arguments>)

Given a PEtab-select YAML file perform model selection with the algorithms specified in the YAML file.

Results are written to a YAML file in the same directory as the PEtab-select YAML file.

Each candidate model produced during the model selection undergoes parameter estimation using local multi-start optimization. Three alg are supported: optimizer=Fides() (Fides Newton-trust region), optimizer=IPNewton() from Optim.jl, and optimizer=LBFGS() from Optim.jl. Additional keywords for the optimisation are n_multistarts::Int- number of multi-starts for parameter estimation (defaults to 100) and optimizationSamplingMethod - which is any sampling method from QuasiMonteCarlo.jl for generating start guesses (defaults to LatinHypercubeSample).

Simulation options can be set using any keyword argument accepted by the PEtabODEProblem function. For example, setting gradient_method=:ForwardDiff specifies the use of forward-mode automatic differentiation for gradient computation. If left blank, we automatically select appropriate options based on the size of the problem.

Note

To use Optim optimizers, you must load Optim with using Optim. To use Ipopt, you must load Ipopt with using Ipopt. To use Fides, load PyCall with using PyCall and ensure Fides is installed (see documentation for setup).

PEtab.solve_SBML — Function

solveSBML(pathSBML, solver, tspan; abstol=1e-8, reltol=1e-8, saveat=Float64[], verbose=true)

Solve an ODE SBML model at the values reported in the SBML file over the specified time span (t0::Float, tend::Float).

Solvers from the OrdinaryDiffEq.jl package are supported. If you want to save the ODE solution at specific time-points, e.g., [1.0, 3.0], provide the saveat argument as saveat=[1.0, 3.0]. The output is provided in the format of OrdinaryDiffEq.jl. The Julia model files are saved in the same directory as the SBML file, in a subdirectory named "SBML".

Note

This function is primarily intended for testing the SBML importer.